NBPackage updates - macOS, templates, utilities

+++ nbpackage/src/main/java/org/apache/netbeans/nbpackage/FileUtils.java
@@ -19,11 +19,13 @@
 package org.apache.netbeans.nbpackage;
 
 import java.io.IOException;
+import java.net.URI;
 import java.nio.file.FileAlreadyExistsException;
 import java.nio.file.FileSystems;
 import java.nio.file.FileVisitResult;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.nio.file.PathMatcher;
 import java.nio.file.SimpleFileVisitor;
 import java.nio.file.StandardCopyOption;
 import java.nio.file.attribute.BasicFileAttributes;
@@ -31,6 +33,7 @@ import java.nio.file.attribute.DosFileAttributeView;
 import java.nio.file.attribute.PosixFileAttributeView;
 import java.nio.file.attribute.PosixFilePermission;
 import java.util.List;
+import java.util.Map;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import org.apache.commons.compress.archivers.ArchiveException;
@@ -141,6 +144,9 @@ public class FileUtils {
             @Override
             public FileVisitResult preVisitDirectory(Path dir, BasicFileAttributes attrs)
                     throws IOException {
+                if (dir.equals(src)) {
+                    return CONTINUE;
+                }
                 Path targetDir = dst.resolve(src.relativize(dir));
                 try {
                     Files.copy(dir, targetDir);
@@ -165,11 +171,59 @@ public class FileUtils {
             @Override
             public FileVisitResult postVisitDirectory(Path dir, IOException exc)
                     throws IOException {
+                if (dir.equals(src)) {
+                    return CONTINUE;
+                }
                 Files.delete(dir);
                 return CONTINUE;
             }
 
         });
+    }
+
+    /**
+     * Recursively delete a directory and all files inside it.
+     *
+     * @param dir directory to delete
+     * @throws IOException
+     */
+    public static void deleteFiles(Path dir) throws IOException {
+        Files.walkFileTree(dir, new SimpleFileVisitor<Path>() {
+
+            @Override
+            public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) throws IOException {
+                Files.delete(file);
+                return FileVisitResult.CONTINUE;
+            }
+
+            @Override
+            public FileVisitResult postVisitDirectory(Path dir, IOException exc) throws IOException {
+                Files.delete(dir);
+                return FileVisitResult.CONTINUE;
+            }
+
+        });
+    }
+
+    /**
+     * Find all files that match the provided glob pattern within a directory.
+     * Files will be included if their relative path within the directory, or
+     * their filename, match the pattern.
+     *
+     * @param searchDir directory to search within
+     * @param pattern pattern to search for
+     * @return list of files that match pattern
+     * @throws IOException
+     */
+    public static List<Path> find(Path searchDir, String pattern) throws IOException {
+        var matcher = searchDir.getFileSystem().getPathMatcher("glob:" + pattern);
+        try (var stream = Files.find(searchDir, Integer.MAX_VALUE, (file, attr)
+                -> !file.equals(searchDir) && 
+                        (matcher.matches(searchDir.relativize(file)) ||
+                                matcher.matches(file.getFileName()))
+        )) {
+            return stream.collect(Collectors.toList());
+        }
     }
 
     /**
@@ -204,6 +258,77 @@ public class FileUtils {
         })) {
             return List.copyOf(stream.collect(Collectors.toList()));
         }
+    }
+
+    /**
+     * Process files inside a JAR file. This can be used, for example, to sign
+     * native executables inside a JAR. Files whose path inside the JAR matches
+     * the given pattern will be extracted into a temporary directory and passed
+     * to the provided processor. If the processor returns true the file will be
+     * updated inside the JAR.
+     *
+     * @param jarFile JAR file
+     * @param pattern glob pattern to match paths for processing
+     * @param processor process files
+     * @return true if the JAR contents have been updated
+     * @throws IOException
+     */
+    public static boolean processJarContents(Path jarFile, String pattern, JarProcessor processor)
+            throws IOException {
+        URI jarURI = URI.create("jar:" + jarFile.toUri());
+        boolean updated = false;
+        try (var jarFS = FileSystems.newFileSystem(jarURI, Map.of())) {
+            PathMatcher matcher = jarFS.getPathMatcher("glob:" + pattern);
+            List<Path> filesToProcess;
+            try (var jarStream = Files.walk(jarFS.getPath("/"))) {
+                filesToProcess = jarStream
+                        .filter(file -> Files.isRegularFile(file))
+                        .filter(matcher::matches)
+                        .collect(Collectors.toList());
+            }
+            if (filesToProcess.isEmpty()) {
+                return false;
+            }
+            Path tmpDir = Files.createTempDirectory("nbpackage-jar-");
+            try {
+                for (Path file : filesToProcess) {
+                    Path tmp = null;
+                    try {
+                        tmp = Files.copy(file, tmpDir.resolve(file.getFileName().toString()));
+                        boolean processed = processor.processFile(tmp, file.toString());
+                        if (processed) {
+                            Files.copy(tmp, file, StandardCopyOption.REPLACE_EXISTING);
+                            updated = true;
+                        }
+                    } finally {
+                        Files.deleteIfExists(tmp);
+                    }
+                }
+
+            } finally {
+                Files.delete(tmpDir);
+            }
+        }
+        return updated;
+    }
+
+    /**
+     * Processor for updating files inside a JAR. For use with
+     * {@link FileUtils#processJarContents(java.nio.file.Path, java.lang.String, org.apache.netbeans.nbpackage.FileUtils.JarProcessor)}.
+     */
+    @FunctionalInterface
+    public static interface JarProcessor {
+
+        /**
+         * Process file from JAR.
+         *
+         * @param tmpFile path to temporary extracted file
+         * @param jarPath full path inside JAR
+         * @return true to update the file in the JAR
+         * @throws IOException
+         */
+        public boolean processFile(Path tmpFile, String jarPath) throws IOException;
+
     }
 
     /**
